deployment/huggingface/requirements.txt
gradio>=4.44.0
ultralytics>=8.3.0
Pillow>=10.0.0
numpy>=1.24.0
torch>=2.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
opencv-python-headless>=4.8.0